Made miscellaneous options script and monitor option
If my monitor wasn't plugged in, but I plug it in, it doesn't
necessarily change my xrandr and stuff.

I have a script to do that on startup, but it only does that if it was
on start up.

Was not comfortable putting undervolt and monitor next to Shutdown and
Restart, now I have computer.sh for "dangerous" commands and a
miscellanous.sh for other stuff.